Created Narration and scene transistions.prefab which contains the initial narration trigger and the objects associated with transitioning to the main scene. Just add in the prefab to the menu scene and line up InitialNarrative with the player and the cylinder and box so that fits in the doorway. These can of course be positioned at will. Narration will not be played in pause menu. This is because of NarrationOnce.cs, which destroys all objects with the DestroyOnPause tag if the "PAUSING" flag is set to one in PlayerPrefs.
Note: In order to prevent overriding other people's stuff in the scene, I have created prefabs for the stuff I made. (SceneFadeIn for Main Menu and aforementioned)
I've also made a prefab of the world lighting in the main scene for use in the menu scene. Prefab is called World Lighting.prefab.