Map finished
Added the screenshot of the environment to the map texture on the
scroll. made it a little scruffy for aesthetic effect. May have copied
and pasted a folder into a certain different area but everything seems
to work fine as i copied it back. All the other assets that go on top of
the map are there and to scale for the area they will be in, should just
be a case of coding them to hover over the area.