Map alterations
there are 2 new materials and 2 new textures. Scroll_texture_2 has the
entrance and exit rune/stones. Scroll_texture_3 is just the square.
Click and drag the second scroll and pick and change the materials as
you please. (i would put them in but its not letting me drag into the
scene for some reason. or navigate for that matter)