Added all my changes that weren't in Show-Peter-Build (listed below) Added in Roy's hand and boxyman model Added a couple of random mini waterfalls (waterfell?!) in the scene, just to pretty it a little Circumvented a bug (the first gate wasnt opening, or at least the mesh wasn't, so no longer displaying mesh)
My changelog:
Visual feedback for the puzzle. Now when a player presses an obelisk, a relic is placed in the obelisk. The order of the obelisks corresponds to the order of relic-picky-uppy.
Character body for better player shadow (immersion and stuffs)

The actually visual feedback could be tweaked, by say Hattle or whatever. By that I just mean the colours of the lights and relic materials

Signed-off-by: Sean Vieira <1000074@live.abertay.ac.uk>